update:change entry point of js
