+++ frontEnd/public/index.html
@@ -8,7 +8,7 @@
 
     <script src="https://kit.fontawesome.com/53b577849a.js" crossorigin="anonymous"></script>
     <link rel="stylesheet" href="./style/style.css">
-    <script defer type="module" src="./js/chatBot.js"></script>
+    <script defer type="module" src="./js/init.js"></script>
 </head>
 <body>
     <section class="navbar">
