+++ frontEnd/public/js/init.js
@@ -1,0 +1,3 @@
+import { eventToBtns } from "./main/component/button.js ";
+
+eventToBtns();
